+++ embedded-io/src/impls/boxx.rs
@@ -1,8 +1,8 @@
-use crate::{BufRead, Io, Read, ReadReady, Seek, Write, WriteReady};
+use crate::{BufRead, ErrorType, Read, ReadReady, Seek, Write, WriteReady};
 use alloc::boxed::Box;
 
 #[cfg_attr(docsrs, doc(cfg(any(feature = "std", feature = "alloc"))))]
-impl<T: ?Sized + Io> Io for Box<T> {
+impl<T: ?Sized + ErrorType> ErrorType for Box<T> {
     type Error = T::Error;
 }